+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [2026-03-03]
+
+### Added
+
+- Notes can be input without an octave. All valid octave combinations within the instrument's range are tried and displayed in the score.
+
 ## [2025-08-28]
 
 ### Added